requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-socketio==5.3.6
python-socketio==5.11.2
eventlet==0.36.1
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
werkzeug==3.0.4